+++ packages/examples/demo_switch/config.tsx
@@ -1,28 +1,17 @@
 import React from "react";
 import {
-  Utils as QbUtils, 
-  Widgets, Fields, Config, Settings, SpelConcatPart, WidgetProps
+  BasicFuncs, Widgets, Fields, Config, Settings, Funcs,
 } from "@react-awesome-query-builder/ui";
 import { MuiConfig } from "@react-awesome-query-builder/mui";
-import ReactSelect from "./select";
+import { getCaseValueWidgetConfig } from "./spel_concat";
+import InputAdornment from "@mui/material/InputAdornment";
 
 export default (): Config => {
   const InitialConfig = MuiConfig;
 
   const widgets: Widgets = {
     ...InitialConfig.widgets,
-    case_value: {
-      ...InitialConfig.widgets.case_value,
-      spelFormatValue: QbUtils.ExportUtils.spelFormatConcat,
-      spelImportValue: QbUtils.ExportUtils.spelImportConcat,
-      jsonLogic: QbUtils.ExportUtils.jsonLogicFormatConcat,
-      factory: ({value, setValue, id}: WidgetProps) => 
-        <ReactSelect 
-          value={value as SpelConcatPart[]}
-          setValue={setValue}
-          k={id!}
-        />
-    }
+    case_value: getCaseValueWidgetConfig(InitialConfig),
   };
 
   const fields: Fields = {
@@ -35,11 +24,36 @@ export default (): Config => {
     qty: {
       label: "Qty",
       type: "number",
+      valueSources: ["value"],
       fieldSettings: {
-        min: 0
+        min: 0,
+        max: 1000,
       },
-      valueSources: ["value", "field"],
       preferWidgets: ["number"]
+    },
+    discount: {
+      label: "Discount",
+      type: "number",
+      valueSources: ["value"],
+      fieldSettings: {
+        min: 0,
+        max: 100,
+        step: 5,
+      },
+      preferWidgets: ["slider", "rangeslider"],
+      widgets: {
+        slider: {
+          widgetProps: {
+            customProps: {
+              input: {
+                InputProps: {
+                  endAdornment: <InputAdornment position="end">%</InputAdornment>
+                }
+              }
+            }
+          }
+        }
+      }
     },
     price: {
       label: "Price",
@@ -59,29 +73,63 @@ export default (): Config => {
       fieldSettings: {
         listValues: [
           { value: "yellow", title: "Yellow" },
+          { value: "blue", title: "Blue" },
           { value: "green", title: "Green" },
           { value: "orange", title: "Orange" }
         ]
       }
     },
     is_promotion: {
-      label: "Promo?",
+      label: "Promotoion",
       type: "boolean",
       operators: ["equal"],
-      valueSources: ["value"]
-    }
+      valueSources: ["value"],
+      mainWidgetProps: {
+        hideOperator: true,
+        operatorInlineLabel: "is",
+      }
+    },
   };
 
   const settings: Settings = {
     ...InitialConfig.settings,
-    maxNumberOfCases: 3,
+    caseValueField: {
+      // type: "case_value", // >> Uncomment to see deprecated using of SpelConcatPart ("./spel_concat.tsx")
+      type: "select",
+      valueSources: ["value"],
+      fieldSettings: {
+        listValues: [{
+          value: "Hot",
+          title: "Hot"
+        }, {
+          value: "In stock",
+          title: "In stock"
+        }, {
+          value: "other",
+          title: "other"
+        }],
+      },
+      mainWidgetProps: {
+        valueLabel: "Then",
+        valuePlaceholder: "Then",
+      },
+    },
+    maxNumberOfCases: 5,
     canRegroupCases: true,
     maxNesting: 3,
     canLeaveEmptyCase: false,
+    showErrorMessage: true,
+  };
+
+  const funcs: Funcs = {
+    LOWER: BasicFuncs.LOWER,
+    UPPER: BasicFuncs.UPPER,
+    LINEAR_REGRESSION: BasicFuncs.LINEAR_REGRESSION,
   };
 
   const config: Config = {
     ...InitialConfig,
+    funcs,
     widgets,
     fields,
     settings