+++ packages/tests/karma.tests.js
@@ -1,7 +1,7 @@
 import Enzyme from "enzyme";
 import Adapter from "@wojtekmaj/enzyme-adapter-react-17";
 import {setCurrentTest, setCurrentTestName, setFilterSpec, getCurrentTestName} from "./support/utils";
-import '@react-awesome-query-builder/ui/css/styles.scss';
+import "@react-awesome-query-builder/ui/css/styles.scss";
 Enzyme.configure({adapter: new Adapter()});
 
 // FILTER YOUR TESTS HERE
@@ -11,6 +11,7 @@ const specFilter = [
 ];
 
 const origDescribe = describe;
+// eslint-disable-next-line no-global-assign
 describe = function (suiteName, fn) {
   setCurrentTestName(suiteName);
   return origDescribe.call(this, suiteName, fn);